requirements.txt
attrs==19.3.0
click==7.1.2
coverage==5.2.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.3
python-dotenv==0.13.0
six==1.15.0
wcwidth==0.2.4
Werkzeug==0.16.0
better-profanity==0.6.1
password-validator==1.0
flask-wtf==0.14.2
WTForms~=2.3.3
fuzzywuzzy~=0.18.0
colorama~=0.4.3
pip~=19.0.3
atomicwrites~=1.4.0
setuptools~=40.8.0
zipp~=3.1.0
